+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@athenna/http",
-  "version": "3.8.0",
+  "version": "3.9.0",
   "description": "The Athenna Http server. Built on top of fastify.",
   "license": "MIT",
   "author": "João Lenon <lenon@athenna.io>",
@@ -61,38 +61,22 @@
     "#tests": "./tests/index.js"
   },
   "dependencies": {
-    "fastify": "^4.14.0"
+    "fastify": "^4.15.0"
   },
   "devDependencies": {
-    "@athenna/artisan": "^3.7.0",
-    "@athenna/common": "^3.5.0",
-    "@athenna/config": "^3.3.0",
-    "@athenna/ioc": "^3.2.0",
-    "@athenna/logger": "^3.2.0",
-    "@athenna/test": "^3.3.0",
-    "@athenna/view": "^3.1.0",
+    "@athenna/artisan": "^3.8.0",
+    "@athenna/common": "^3.5.1",
+    "@athenna/config": "^3.4.1",
+    "@athenna/ioc": "^3.3.0",
+    "@athenna/logger": "^3.3.0",
+    "@athenna/test": "^3.5.1",
+    "@athenna/view": "^3.2.0",
     "@fastify/cors": "^8.1.1",
     "@fastify/helmet": "^10.0.2",
     "@fastify/rate-limit": "^7.5.0",
     "@fastify/swagger": "^8.1.0",
     "@fastify/swagger-ui": "^1.1.0",
-    "@japa/assert": "^1.3.6",
-    "@japa/run-failed-tests": "^1.1.0",
-    "@japa/runner": "^2.2.2",
-    "@japa/spec-reporter": "^1.3.2",
     "@swc/core": "^1.3.27",
-    "@types/bytes": "^3.1.1",
-    "@types/callsite": "^1.0.31",
-    "@types/columnify": "^1.5.1",
-    "@types/debug": "^4.1.7",
-    "@types/figlet": "^1.5.5",
-    "@types/inquirer": "^9.0.3",
-    "@types/kind-of": "^6.0.0",
-    "@types/lodash": "^4.14.191",
-    "@types/ms": "^0.7.31",
-    "@types/pluralize": "^0.0.29",
-    "@types/sinon": "^10.0.13",
-    "@types/uuid": "^9.0.0",
     "@typescript-eslint/eslint-plugin": "^5.56.0",
     "@typescript-eslint/parser": "^5.56.0",
     "c8": "^7.12.0",
@@ -108,12 +92,9 @@
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-promise": "^6.1.1",
     "husky": "^3.1.0",
-    "js-yaml": "^4.1.0",
     "lint-staged": "^12.5.0",
-    "minimist": "^1.2.7",
     "prettier": "^2.8.7",
     "reflect-metadata": "^0.1.13",
-    "sinon": "^15.0.2",
     "ts-node": "^10.9.1",
     "typescript": "^5.0.2"
   },
@@ -171,9 +152,7 @@
       }
     ]
   },
-  "eslintIgnore": [
-    "build/**/*"
-  ],
+  "eslintIgnore": [],
   "eslintConfig": {
     "env": {
       "es2021": true,
